Extract tests to separate file
